fuse 0.2.2

+++ src/day15.rs
@@ -100,7 +100,7 @@ impl Filesystem for JsonFilesystem {
         };
     }
 
-    fn read(&mut self, _req: &Request, ino: u64, fh: u64, offset: u64, size: usize, reply: ReplyData) {
+    fn read(&mut self, _req: &Request, ino: u64, fh: u64, offset: u64, size: u32, reply: ReplyData) {
         println!("read(ino={}, fh={}, offset={}, size={})", ino, fh, offset, size);
         for (key, &inode) in self.inodes.iter() {
             if inode == ino {
